Speculate to avoid testing equivalent conditions
Use Speculate to avoid testing equivalent conditions.  Now, for each run of
Extrapolate where there is a fault, there is a small price payd for
Speculating.  The benefits are:

* conditions are smaller: "not (x <= y)" -> "x > y";
* encoded assignments are detected and not reported: "not (x /= y)";
* examples that were very slow had their runtime reduced;

The downside:

* examples that were really fast had their runtime increased.

Note that "bench/false" and "bench/ord" had their runtime increased a lot since
they are a set of small examples and each example takes a few seconds to run.